fix: Resolve CI failures - clippy and security issues
- Remove unnecessary type cast in scanner.rs (i64 -> i64)
- Update ring dependency from 0.17.9 to 0.17.14 to fix RUSTSEC-2025-0009
- Fixes GitHub Actions clippy check failure
- Addresses security vulnerability in AES functions

+++ src/scanner.rs
@@ -1745,7 +1745,7 @@ impl Scanner {
             return FilesystemType::Mem8;
         }
 
-        match stat_buf.f_type as i64 {
+        match stat_buf.f_type {
             EXT4_SUPER_MAGIC => FilesystemType::Ext4, // TODO: Distinguish ext2/3/4
             XFS_SUPER_MAGIC => FilesystemType::Xfs,
             BTRFS_SUPER_MAGIC => FilesystemType::Btrfs,
